feat(svg): move SVG sanitization tests to a dedicated file for better organization
